renamed xarray-healpy to xhealpixify.
